Market at home

+++ app/src/main/java/com/doubleclick/marktinhome/ui/MainScreen/Frgments/Add/UploadFragment.kt
@@ -85,8 +85,8 @@ class UploadFragment : BaseFragment() {
         Upload.setOnClickListener {
             upload(
                 productName.text.toString(),
-                productPrice.text.toString(),
-                productLastPrice.text.toString(),
+                productPrice.text.toString().toDouble(),
+                productLastPrice.text.toString().toDouble(),
                 description.text.toString(),
                 keywords.text.toString(),
                 marke,
